New script added

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int i, n = 50;
-	int fib1 = 1, fib2 = 2, fib3;
+	long  fib1 = 1, fib2 = 2, fib3;
 
-	printf("%d, %d, ", fib1, fib2);
+	printf("%lu, %lu, ", fib1, fib2);
 
 	for (i = 3; i <= n; i++)
 	{
 		fib3 = fib1 + fib2;
-		printf("%d", fib3);
+		printf("%lu", fib3);
 		fib1 = fib2;
 		fib2 = fib3;
